correct scaling of player+minimap size check before draw_walls

+++ src/init/init_camera.c
@@ -6,7 +6,7 @@
 /*   By: vmanuyko <vmanuyko@student.42vienna.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/21 11:50:22 by vmanuyko          #+#    #+#             */
-/*   Updated: 2026/03/21 13:56:39 by vmanuyko         ###   ########.fr       */
+/*   Updated: 2026/03/21 16:11:33 by vmanuyko         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ void	init_camera(t_camera *cam, t_user *user)
 
 	tiles = (MINI_SIZE / MINI_TILE);
 	cam->min.x = user->player.pos.x - tiles / 2;
-	cam->max.x = (int)cam->min.x + tiles;
+	cam->max.x = cam->min.x + tiles;
 	cam->min.y = user->player.pos.y - tiles / 2;
-	cam->max.y = (int)cam->min.y + tiles;
+	cam->max.y = cam->min.y + tiles;
 }
